+++ insta_post/views.py
@@ -10,10 +10,10 @@ from insta_post.forms import PostForm
 from insta_comment.forms import CommentForm
 # from insta_comment.helpers import add_one
 
+
 class IndexView(TemplateView):
   
     def get(self,request):
-        # Do not change this objects.all!!!! Rubric item!!!
         cars = FavoriteCar.objects.all()
         return render(request, "index.html", {"cars": cars})
 
@@ -87,24 +87,6 @@ def del_post(request, post_id):
     else: 
         return HttpResponseForbidden("You do not have permission to delete this post")
      
-
-# @login_required
-class FollowView(TemplateView):
-    def get(self, request, follow_id):
-        signed_in_user = InstaUser.objects.filter(username=request.user.username).first()
-        follow = InstaUser.objects.filter(id=follow_id).first()
-        if follow_id != request.user.id:
-            signed_in_user.following.add(follow)
-        return HttpResponseRedirect(request.META.get('HTTP_REFERER', '/'))
-
-# @login_required
-class UnfollowView(TemplateView):
-    def get(self, request, unfollow_id):
-        signed_in_user = request.user
-        unfollow = InstaUser.objects.filter(id=unfollow_id).first()
-        signed_in_user.following.remove(unfollow)
-        return HttpResponseRedirect(request.META.get('HTTP_REFERER', '/'))
-
 
 def handler404(request, exception):
     return render(request, '404.html', status=404)
