+++ static/css/lori-style.css
@@ -1,5 +1,5 @@
 * {
-    text-align: center;
+  text-align: center;
 }
 
 background-image: -webkit-repeating-radial-gradient(blue, yellow);
